+++ distribution/galleon-feature-packs/server-galleon-pack/src/main/resources/modules/system/layers/keycloak/org/keycloak/keycloak-model-legacy-services/main/module.xml
@@ -39,6 +39,9 @@
         <module name="org.keycloak.keycloak-services" services="import"/>
         <module name="org.keycloak.keycloak-wildfly-extensions" export="true" services="import"/>
 
+        <!-- Openshift Client Storage -->
+        <module name="com.openshift.openshift-restclient-java" services="import"/>
+
         <module name="com.googlecode.owasp-java-html-sanitizer"/>
         <module name="com.google.guava"/>
         <module name="org.freemarker"/>
