+++ model/legacy-private/src/main/java/org/keycloak/storage/client/AbstractClientStorageAdapter.java
