+++ model/legacy-private/src/main/java/org/keycloak/storage/client/AbstractReadOnlyClientStorageAdapter.java
