+++ model/legacy-private/src/main/resources/META-INF/services/org.keycloak.provider.Spi
@@ -17,3 +17,4 @@
 
 org.keycloak.models.cache.CacheUserProviderSpi
 org.keycloak.models.cache.CacheRealmProviderSpi
+org.keycloak.storage.client.ClientStorageProviderSpi
