+++ model/legacy-services/src/main/java/org/keycloak/services/resources/admin/ClientStorageProviderResource.java
