+++ model/legacy-services/src/main/resources/META-INF/services/org.keycloak.storage.client.ClientStorageProviderFactory
