+++ server-spi-private/src/main/resources/META-INF/services/org.keycloak.provider.Spi
@@ -78,7 +78,6 @@ org.keycloak.credential.CredentialSpi
 org.keycloak.keys.PublicKeyStorageSpi
 org.keycloak.keys.KeySpi
 org.keycloak.storage.DatastoreSpi
-org.keycloak.storage.client.ClientStorageProviderSpi
 org.keycloak.storage.clientscope.ClientScopeStorageProviderSpi
 org.keycloak.storage.role.RoleStorageProviderSpi
 org.keycloak.storage.group.GroupStorageProviderSpi
