+++ services/src/main/java/org/keycloak/social/facebook/FacebookIdentityProviderConfig.java
@@ -2,10 +2,8 @@ package org.keycloak.social.facebook;
 
 import java.util.Optional;
 
-import org.apache.commons.lang.StringUtils;
 import org.keycloak.broker.oidc.OIDCIdentityProviderConfig;
 import org.keycloak.models.IdentityProviderModel;
-import org.keycloak.saml.common.util.StringUtil;
 
 public class FacebookIdentityProviderConfig extends OIDCIdentityProviderConfig {
