+++ services/src/test/java/org/keycloak/connections/httpclient/DefaultHttpClientFactoryTest.java
@@ -30,7 +30,6 @@ import java.util.Properties;
 
 import javax.net.ssl.SSLPeerUnverifiedException;
 
-import org.apache.commons.lang.StringUtils;
 import org.apache.http.HttpStatus;
 import org.apache.http.client.methods.CloseableHttpResponse;
 import org.apache.http.client.methods.HttpGet;
@@ -95,7 +94,7 @@ public class DefaultHttpClientFactoryTest {
 
 		StringBuilder sb = new StringBuilder();
 		sb.append("{");
-		sb.append(StringUtils.join(params, ','));
+		sb.append(String.join(",", params));
 		sb.append("}");
 		
 		return sb.toString();
